Cleanup events (partially) (#96)
