+++ src/index.ts
@@ -9,7 +9,7 @@ import * as phaser from 'phaser-ce';
 import * as common from './common';
 import * as events from './events';
 import * as generators from './events/generators';
-
+import * as textures from './character/textures';
 import Boot from './ui/states/boot';
 import { generateMap } from './map/generator';
 import Main from './ui/states/main';
@@ -17,8 +17,10 @@ import WorldState from './world_state/world_state';
 import HudModel from './ui/hud/hud_model';
 import { ITicker } from './ticker';
 import Controller from './input/controller';
-import { SpawnConfig } from './character/spawn_config';
+import { SpawnConfig, randomSpawnLocation } from './character/spawn_config';
 import { Armory } from './ui/sprites/armory';
+import Character from './character/character';
+import HudBuilder from './ui/hud/hud_builder';
 
 export class Game extends phaser.Game {
   public generators: ITicker[];
@@ -27,6 +29,7 @@ export class Game extends phaser.Game {
   public worldState: WorldState;
   public controller: Controller;
   public readonly armory: Armory;
+  public isOfferingContract: boolean;
 
   constructor() {
     super({
@@ -41,6 +44,8 @@ export class Game extends phaser.Game {
     this.state.add('Main', Main);
     this.state.start('Boot');
 
+    this.isOfferingContract = false;
+    this.hud = new HudBuilder().build();
     this.armory = new Armory(this);
     this.worldState = new WorldState(40, 40);
 
@@ -61,8 +66,36 @@ export class Game extends phaser.Game {
   }
 
   public spawn(config: SpawnConfig) {
-    const e = new events.Event(events.EventType.CharacterSpawn, config, 0);
-    this.gameEvents.emit(e.type, e);
+    if (common.experiment('spawn')) {
+      const sprite = game.add.sprite(
+        config.x * 64,
+        config.y * 64,
+        config.texture
+      );
+      sprite.scale = new Phaser.Point(4.0, 4.0);
+      game.physics.p2.enable(sprite);
+      config.character.setSprite(sprite);
+      game.worldState.characters.push(config.character);
+    }
+  }
+
+  public offerContract(character: Character) {
+    this.isOfferingContract = true;
+    const { x, y } = randomSpawnLocation(game.worldState.grid);
+    this.hud = this.hud.setQuestion(
+      'Do you want to hire ' + character.name + '?',
+      ['yes', 'no'],
+      (option: number) => {
+        this.isOfferingContract = false;
+        this.hud = this.hud.clearQuestion();
+        if (option === 1) {
+          this.hud = this.hud.setMessage('You hired ' + character.name + '!');
+          this.spawn(
+            new SpawnConfig(character, textures.guard(game.armory), x, y)
+          );
+        }
+      }
+    );
   }
 }
